% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ops.R
\name{geneFromCache}
\alias{geneFromCache}
\title{download a parquet file from the OSN bucket to local BiocFileCache}
\usage{
geneFromCache(resource, cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{resource}{character(1)}

\item{cache}{BiocFileCache object}
}
\value{
path to local cached file
}
\description{
download a parquet file from the OSN bucket to local BiocFileCache
}
