% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ops.R
\name{is_online}
\alias{is_online}
\title{utility to check for internet access for testing}
\usage{
is_online(site = "http://example.com/")
}
\arguments{
\item{site}{character(1)}
}
\value{
logical(1)
}
\description{
utility to check for internet access for testing
}
