% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGparq.R
\name{NGparq}
\alias{NGparq}
\title{obtain references to parquet serializations of slices
of NCBI Gene annotation}
\usage{
NGparq()
}
\description{
obtain references to parquet serializations of slices
of NCBI Gene annotation
}
\examples{
ng = NGparq()
sapply(ng, nrow)
}
