% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGparq.R
\name{print.NGparq}
\alias{print.NGparq}
\title{print method for NGparq list}
\usage{
\method{print}{NGparq}(x, ...)
}
\arguments{
\item{x}{instance of S3 class NGparq}

\item{\dots}{not used}
}
\description{
print method for NGparq list
}
