% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGparq.R
\name{processDbx}
\alias{processDbx}
\title{given a bar-delimited string as produced for dbXrefs in gene_info,
extract the value associated with a given tag}
\usage{
processDbx(x, tag = "Ensembl:")
}
\arguments{
\item{x}{character(1)}

\item{tag}{substring to use to find desired value}
}
\description{
given a bar-delimited string as produced for dbXrefs in gene_info,
extract the value associated with a given tag
}
\examples{
dem = "MIM:611221|HGNC:HGNC:23690|Ensembl:ENSG00000073605|AllianceGenome:HGNC:23690"
processDbx(dem, tag="MIM:")
}
