% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGparq.R
\name{processDbx1}
\alias{processDbx1}
\title{given an element of a bar-delimited string as produced for dbXrefs in gene_info,
extract the value associated with a given tag}
\usage{
processDbx1(x, tag = "Ensembl:")
}
\arguments{
\item{x}{character(1)}

\item{tag}{substring to use to find desired value}
}
\description{
given an element of a bar-delimited string as produced for dbXrefs in gene_info,
extract the value associated with a given tag
}
