% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ops.R
\name{taxon_cache_info}
\alias{taxon_cache_info}
\title{list taxon-specific parquets stored in BiocFileCache}
\usage{
taxon_cache_info(taxid = NULL, cache = .ncbi_cache())
}
\arguments{
\item{taxid}{integer(1) or NULL; when NULL lists all cached taxon parquets}

\item{cache}{a \code{BiocFileCache} object}
}
\value{
data.frame with columns \code{rname}, \code{rpath}, \code{create_time}
}
\description{
list taxon-specific parquets stored in BiocFileCache
}
\examples{
taxon_cache_info()
}
