% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomyMap.R
\name{taxonomyMap}
\alias{taxonomyMap}
\title{retrieve a vector of taxonomy names indexed by taxonomy codes}
\usage{
taxonomyMap()
}
\description{
retrieve a vector of taxonomy names indexed by taxonomy codes
}
\note{
retrieved from https://ftp.ncbi.nlm.nih.gov/pub/taxonomy/new_taxdump/taxdump_readme.txt
Feb 22 2025
}
\examples{
tmap = taxonomyMap()
tmap['9606']
}
