% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ops.R
\name{geneFromCache}
\alias{geneFromCache}
\title{populate cache with available parquet files if needed, return
path to cached file}
\usage{
geneFromCache(resource, cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{resource}{character(1)}

\item{cache}{character(1) BiocFileCache-like object}
}
\value{
path to local version of resource
}
\description{
populate cache with available parquet files if needed, return
path to cached file
}
\examples{
gi = geneFromCache("gene_info.parquet")
arrow::open_dataset(gi) |> dplyr::filter(`#tax_id`==9606) |> head() |> dplyr::collect()
}
