% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIdsNG.R
\name{mapIdsNG}
\alias{mapIdsNG}
\title{very preliminary analog of mapIds from AnnotationDbi, uses remote query to parquet
for NCBI gene_info; the NG connotes NCBI Gene}
\usage{
mapIdsNG(
  taxid = 9606L,
  keys = c("ORMDL3", "TP53", "GSDMB", "XyZZY"),
  keytype = "Symbol",
  columns = c("GeneID", "type_of_gene", "description")
)
}
\arguments{
\item{taxid}{integer(1) integer code for species, defaults to 9606L for Homo sapiens}

\item{keys}{a vector of keys to be translated}

\item{keytype}{character(1), at present, some column of the gene_info.parquet dataset}

\item{columns}{character(), a subset of columns of the gene_info.parquet dataset}
}
\description{
very preliminary analog of mapIds from AnnotationDbi, uses remote query to parquet
for NCBI gene_info; the NG connotes NCBI Gene
}
\note{
At present this function only uses `remote_gene_query`
}
\examples{
if (interactive()) {
mapIdsNG()
}
}
