% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{ncbi_gene_con}
\alias{ncbi_gene_con}
\title{get or create a persistent duckdb connection with httpfs loaded}
\usage{
ncbi_gene_con()
}
\value{
a DBI connection to an in-process duckdb instance
}
\description{
get or create a persistent duckdb connection with httpfs loaded
}
