% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{remote_gene_query}
\alias{remote_gene_query}
\title{use duckdb to query NCBI Gene data in OSN bucket}
\usage{
remote_gene_query(
  gres = "gene2pubmed",
  qual = "limit 5",
  tname = basename(tempfile())
)
}
\arguments{
\item{gres}{name of a gene resource, no suffix, see available_gene_parquet vector (unexported)}

\item{qual}{a SQL fragment used to qualify a select * clause}

\item{tname}{character(1) arbitrary name to use for internal sql table}
}
\description{
use duckdb to query NCBI Gene data in OSN bucket
}
\note{
The httpfs extension for duckdb is installed when the function is called.
}
\examples{
if (is_online()) {
remote_gene_query(qual = 'where "#tax_id" = 9606 limit 10;')
}
}
